+++ backend/src/data/database.rs
@@ -9,6 +9,7 @@ use sqlx::{FromRow, QueryBuilder, Sqlite, SqlitePool};
 use tokio::task::spawn_blocking;
 use tracing::{info, warn};
 
+use crate::service::commands::ffmpeg::FfmpegLocation;
 use crate::service::music;
 use crate::Result;
 
@@ -39,6 +40,7 @@ pub struct DbVideo {
     pub interactive: bool,
     pub source: LocalVideoSource,
     pub duration: f64,
+    pub video_preview_image: Option<String>,
 }
 
 #[derive(Debug, Clone, PartialEq, FromRow)]
@@ -50,6 +52,7 @@ pub struct DbMarker {
     pub title: String,
     pub file_path: String,
     pub index_within_video: i64,
+    pub marker_preview_image: Option<String>,
 }
 
 #[derive(Debug, Clone, Deserialize)]
@@ -60,6 +63,7 @@ pub struct CreateMarker {
     pub end: f64,
     pub title: String,
     pub index_within_video: i64,
+    pub preview_image_path: Option<String>,
 }
 
 #[derive(Debug)]
@@ -83,6 +87,12 @@ pub struct CreateSong {
     pub file_path: String,
     pub duration: f64,
     pub beats: Option<Beats>,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub enum AllVideosFilter {
+    NoVideoDuration,
+    NoPreviewImage,
 }
 
 #[derive(Clone)]
@@ -110,7 +120,7 @@ impl Database {
     pub async fn get_video(&self, id: &str) -> Result<Option<DbVideo>> {
         let video = sqlx::query_as!(
             DbVideo,
-            "SELECT id, file_path, interactive, source, duration FROM local_videos WHERE id = $1",
+            "SELECT id, file_path, interactive, source, duration, video_preview_image FROM local_videos WHERE id = $1",
             id
         )
         .fetch_optional(&self.pool)
@@ -135,6 +145,7 @@ impl Database {
                     interactive: group[0].interactive,
                     source: group[0].source.clone().into(),
                     duration: group[0].duration,
+                    video_preview_image: group[0].video_preview_image.clone(),
                 };
                 let markers: Vec<_> = group
                     .into_iter()
@@ -147,6 +158,7 @@ impl Database {
                             r.rowid,
                             r.file_path,
                             r.index_within_video,
+                            r.marker_preview_image,
                         ) {
                             (
                                 Some(video_id),
@@ -156,6 +168,7 @@ impl Database {
                                 rowid,
                                 file_path,
                                 Some(index),
+                                marker_preview_image,
                             ) => Some(DbMarker {
                                 rowid,
                                 title,
@@ -164,6 +177,7 @@ impl Database {
                                 end_time,
                                 file_path,
                                 index_within_video: index,
+                                marker_preview_image,
                             }),
                             _ => None,
                         }
@@ -178,7 +192,7 @@ impl Database {
     pub async fn get_marker(&self, id: i64) -> Result<DbMarker> {
         let marker = sqlx::query_as!(
             DbMarker,
-            "SELECT m.rowid, m.title, m.video_id, v.file_path, m.start_time, m.end_time, m.index_within_video
+            "SELECT m.rowid, m.title, m.video_id, v.file_path, m.start_time, m.end_time, m.index_within_video, m.marker_preview_image
                 FROM markers m INNER JOIN local_videos v ON m.video_id = v.id
                 WHERE m.rowid = $1",
             id
@@ -193,7 +207,7 @@ impl Database {
         ids: &[impl AsRef<str>],
     ) -> Result<Vec<DbMarker>> {
         let mut query_builder: QueryBuilder<Sqlite> = QueryBuilder::new(
-            "SELECT m.rowid, m.title, m.video_id, v.file_path, m.start_time, m.end_time, m.index_within_video
+            "SELECT m.rowid, m.title, m.video_id, v.file_path, m.start_time, m.end_time, m.index_within_video, m.marker_preview_image
             FROM markers m INNER JOIN local_videos v ON m.video_id = v.id
             WHERE m.video_id IN ("
         );
@@ -230,6 +244,7 @@ impl Database {
                 interactive: records[0].interactive,
                 source: records[0].source.clone().into(),
                 duration: records[0].duration,
+                video_preview_image: records[0].video_preview_image.clone(),
             };
             let markers = records
                 .into_iter()
@@ -242,6 +257,7 @@ impl Database {
                         r.rowid,
                         r.file_path,
                         r.index_within_video,
+                        r.marker_preview_image,
                     ) {
                         (
                             Some(video_id),
@@ -251,6 +267,7 @@ impl Database {
                             rowid,
                             file_path,
                             Some(index),
+                            marker_preview_image,
                         ) => Some(DbMarker {
                             rowid,
                             title,
@@ -259,6 +276,7 @@ impl Database {
                             end_time,
                             file_path,
                             index_within_video: index,
+                            marker_preview_image,
                         }),
                         _ => None,
                     }
@@ -268,8 +286,32 @@ impl Database {
         }
     }
 
-    pub async fn get_videos(&self) -> Result<Vec<DbVideo>> {
-        sqlx::query_as!(DbVideo, "SELECT * FROM local_videos")
+    pub async fn get_videos(&self, filter: AllVideosFilter) -> Result<Vec<DbVideo>> {
+        let query = match filter {
+            AllVideosFilter::NoVideoDuration => {
+                sqlx::query_as!(DbVideo, "SELECT * FROM local_videos WHERE duration = -1.0")
+                    .fetch_all(&self.pool)
+                    .await
+            }
+            AllVideosFilter::NoPreviewImage => {
+                sqlx::query_as!(
+                    DbVideo,
+                    "SELECT * FROM local_videos WHERE video_preview_image IS NULL"
+                )
+                .fetch_all(&self.pool)
+                .await
+            }
+        };
+        query.map_err(From::from)
+    }
+
+    pub async fn get_markers_without_preview_images(&self) -> Result<Vec<DbMarker>> {
+        sqlx::query_as!(
+            DbMarker,
+            "SELECT m.rowid, m.title, m.video_id, v.file_path, m.start_time, m.end_time, m.index_within_video, m.marker_preview_image
+            FROM markers m INNER JOIN local_videos v ON m.video_id = v.id
+            WHERE m.marker_preview_image IS NULL"
+        )
             .fetch_all(&self.pool)
             .await
             .map_err(From::from)
@@ -277,12 +319,13 @@ impl Database {
 
     pub async fn persist_video(&self, video: DbVideo) -> Result<()> {
         sqlx::query!(
-            "INSERT INTO local_videos (id, file_path, interactive, source, duration) VALUES ($1, $2, $3, $4, $5)",
+            "INSERT INTO local_videos (id, file_path, interactive, source, duration, video_preview_image) VALUES ($1, $2, $3, $4, $5, $6)",
             video.id,
             video.file_path,
             video.interactive,
             video.source,
             video.duration,
+            video.video_preview_image,
         )
         .execute(&self.pool)
         .await?;
@@ -291,8 +334,8 @@ impl Database {
 
     pub async fn persist_marker(&self, marker: CreateMarker) -> Result<DbMarker> {
         let new_id = sqlx::query_scalar!(
-            "INSERT INTO markers (video_id, start_time, end_time, title, index_within_video) 
-                VALUES ($1, $2, $3, $4, $5) 
+            "INSERT INTO markers (video_id, start_time, end_time, title, index_within_video, marker_preview_image) 
+                VALUES ($1, $2, $3, $4, $5, $6) 
                 ON CONFLICT DO UPDATE SET start_time = excluded.start_time, end_time = excluded.end_time, title = excluded.title
                 RETURNING rowid",
                 marker.video_id,
@@ -300,6 +343,7 @@ impl Database {
                 marker.end,
                 marker.title,
                 marker.index_within_video,
+                marker.preview_image_path,
         )
         .fetch_one(&self.pool)
         .await?;
@@ -312,6 +356,7 @@ impl Database {
             video_id: marker.video_id,
             file_path: "not-needed".to_string(),
             index_within_video: marker.index_within_video,
+            marker_preview_image: marker.preview_image_path,
         };
 
         info!("newly updated or inserted marker: {marker:?}");
@@ -434,7 +479,7 @@ impl Database {
         Ok(())
     }
 
-    pub async fn generate_all_beats(&self) -> Result<()> {
+    pub async fn generate_all_beats(&self, ffmpeg: FfmpegLocation) -> Result<()> {
         let rows = sqlx::query!("SELECT rowid, file_path FROM songs WHERE beats IS NULL")
             .fetch_all(&self.pool)
             .await?;
@@ -444,8 +489,9 @@ impl Database {
         info!("generating beats for {} songs", rows.len());
         let mut handles = vec![];
         for row in rows {
+            let ffmpeg = ffmpeg.clone();
             handles.push(spawn_blocking(move || {
-                (music::detect_beats(row.file_path), row.rowid)
+                (music::detect_beats(row.file_path, &ffmpeg), row.rowid)
             }));
         }
 
@@ -461,6 +507,30 @@ impl Database {
         sqlx::query!(
             "UPDATE local_videos SET duration = $1 WHERE id = $2",
             duration,
+            id
+        )
+        .execute(&self.pool)
+        .await?;
+
+        Ok(())
+    }
+
+    pub async fn set_video_preview_image(&self, id: &str, preview_image: &str) -> Result<()> {
+        sqlx::query!(
+            "UPDATE local_videos SET video_preview_image = $1 WHERE id = $2",
+            preview_image,
+            id
+        )
+        .execute(&self.pool)
+        .await?;
+
+        Ok(())
+    }
+
+    pub async fn set_marker_preview_image(&self, id: i64, preview_image: &str) -> Result<()> {
+        sqlx::query!(
+            "UPDATE markers SET marker_preview_image = $1 WHERE rowid = $2",
+            preview_image,
             id
         )
         .execute(&self.pool)
@@ -501,6 +571,7 @@ mod test {
             start: 0.0,
             end: 17.0,
             index_within_video: 0,
+            preview_image_path: None,
         };
         let result = database.persist_marker(expected.clone()).await.unwrap();
 
@@ -520,6 +591,7 @@ mod test {
             start: 0.0,
             end: 17.0,
             index_within_video: 0,
+            preview_image_path: None,
         };
         let err = database
             .persist_marker(expected.clone())
@@ -540,6 +612,7 @@ mod test {
             start: 0.0,
             end: 17.0,
             index_within_video: 0,
+            preview_image_path: None,
         };
         let result = database.persist_marker(marker).await.unwrap();
         let id = result.rowid.unwrap();