+++ backend/src/server/handlers/local.rs
@@ -10,13 +10,14 @@ use clip_mash_types::{ListVideoDto, MarkerDto, VideoDto};
 use reqwest::StatusCode;
 use serde::Deserialize;
 use tower::ServiceExt;
-use tracing::info;
+use tracing::{info, warn};
 use url::Url;
 
 use crate::data::database::CreateMarker;
 use crate::server::error::AppError;
 use crate::server::handlers::AppState;
 use crate::service::local_video::VideoService;
+use crate::service::preview_image::PreviewGenerator;
 
 #[axum::debug_handler]
 pub async fn get_video(
@@ -29,6 +30,40 @@ pub async fn get_video(
     let video = state.database.get_video(&id).await?;
     if let Some(video) = video {
         let result = ServeFile::new(video.file_path).oneshot(request).await;
+        Ok(result)
+    } else {
+        Err(AppError::StatusCode(StatusCode::NOT_FOUND))
+    }
+}
+
+#[axum::debug_handler]
+pub async fn get_video_preview(
+    Path(id): Path<String>,
+    state: State<Arc<AppState>>,
+    request: axum::http::Request<Body>,
+) -> Result<impl IntoResponse, AppError> {
+    use tower_http::services::ServeFile;
+
+    let video = state.database.get_video(&id).await?;
+    if let Some(preview_image) = video.and_then(|v| v.video_preview_image) {
+        let result = ServeFile::new(preview_image).oneshot(request).await;
+        Ok(result)
+    } else {
+        Err(AppError::StatusCode(StatusCode::NOT_FOUND))
+    }
+}
+
+#[axum::debug_handler]
+pub async fn get_marker_preview(
+    Path(id): Path<i64>,
+    state: State<Arc<AppState>>,
+    request: axum::http::Request<Body>,
+) -> Result<impl IntoResponse, AppError> {
+    use tower_http::services::ServeFile;
+
+    let marker = state.database.get_marker(id).await?;
+    if let Some(preview_image) = marker.marker_preview_image {
+        let result = ServeFile::new(preview_image).oneshot(request).await;
         Ok(result)
     } else {
         Err(AppError::StatusCode(StatusCode::NOT_FOUND))
@@ -84,16 +119,27 @@ fn validate_marker(marker: &CreateMarker) -> HashMap<&'static str, &'static str>
 #[axum::debug_handler]
 pub async fn persist_marker(
     state: State<Arc<AppState>>,
-    Json(marker): Json<CreateMarker>,
+    Json(mut marker): Json<CreateMarker>,
 ) -> Result<Json<MarkerDto>, AppError> {
     let validation = validate_marker(&marker);
     if !validation.is_empty() {
         Err(AppError::Validation(validation))
     } else {
         info!("saving marker {marker:?} to the database");
-        let marker = state.database.persist_marker(marker).await?;
 
-        Ok(Json(marker.into()))
+        if let Some(video) = state.database.get_video(&marker.video_id).await? {
+            let preview_generator: PreviewGenerator = state.0.clone().into();
+            let preview_image = preview_generator
+                .generate_preview(&video.id, &video.file_path, video.duration / 2.0)
+                .await?;
+            marker.preview_image_path = Some(preview_image.to_string());
+            let marker = state.database.persist_marker(marker).await?;
+
+            Ok(Json(marker.into()))
+        } else {
+            warn!("No video found for marker {marker:?}, returning 404");
+            Err(AppError::StatusCode(StatusCode::NOT_FOUND))
+        }
     }
 }