+++ backend/src/service/commands/ffmpeg.rs
@@ -1,8 +1,7 @@
-#![allow(unused)]
-
 use std::fs;
 use std::io::Write;
 use std::process::{Command, Stdio};
+use std::sync::Arc;
 
 use camino::{Utf8Path, Utf8PathBuf};
 use color_eyre::eyre::bail;
@@ -32,25 +31,27 @@ fn download_url() -> Result<(&'static str, &'static str)> {
     }
 }
 
-fn is_installed(dest: &Utf8Path) -> bool {
-    if dest.join("ffmpeg").is_file() || dest.join("ffmpeg.exe").is_file() {
-        true
+fn get_ffmpeg_location(base_dir: &Utf8Path) -> Option<FfmpegLocation> {
+    if base_dir.join("ffmpeg").is_file() || base_dir.join("ffmpeg.exe").is_file() {
+        Some(FfmpegLocation::Local(Arc::new(base_dir.to_owned())))
     } else {
-        Command::new("ffmpeg")
+        let is_system = Command::new("ffmpeg")
             .arg("-version")
             .stderr(Stdio::null())
             .stdout(Stdio::null())
             .status()
             .map(|s| s.success())
-            .unwrap_or_else(|_| false)
+            .unwrap_or_else(|_| false);
+        if is_system {
+            Some(FfmpegLocation::System)
+        } else {
+            None
+        }
     }
 }
 
 #[cfg(target_os = "linux")]
-fn unzip(
-    path: impl AsRef<Utf8Path>,
-    destination_folder: impl AsRef<Utf8Path>,
-) -> Result<Utf8PathBuf> {
+fn unzip(path: impl AsRef<Utf8Path>, destination_folder: impl AsRef<Utf8Path>) -> Result<()> {
     use std::fs::File;
     use std::io::BufReader;
 
@@ -60,26 +61,22 @@ fn unzip(
     let file = BufReader::new(File::open(path.as_ref())?);
     let reader = XzDecoder::new(file);
     let mut archive = Archive::new(reader);
-    let dest_file = destination_folder.as_ref().join("ffmpeg");
-    info!("unzipping to {dest_file}");
 
     for entry in archive.entries()? {
         let mut entry = entry?;
         let path = entry.path()?;
         let path = Utf8Path::from_path(path.as_ref()).unwrap();
-        if let Some("ffmpeg") = path.file_name() {
+        if let Some("ffmpeg") | Some("ffprobe") = path.file_name() {
+            let dest_file = destination_folder.as_ref().join(path.file_name().unwrap());
             entry.unpack(&dest_file)?;
         }
     }
 
-    Ok(dest_file)
+    Ok(())
 }
 
 #[cfg(target_os = "windows")]
-fn unzip(
-    path: impl AsRef<Utf8Path>,
-    destination_folder: impl AsRef<Utf8Path>,
-) -> Result<Utf8PathBuf> {
+fn unzip(path: impl AsRef<Utf8Path>, destination_folder: impl AsRef<Utf8Path>) -> Result<()> {
     use std::fs::File;
     use std::io::BufReader;
 
@@ -87,24 +84,24 @@ fn unzip(
 
     let file = BufReader::new(File::open(path.as_ref())?);
     let mut zip = ZipArchive::new(file)?;
-    let dest_path = destination_folder.as_ref().join("ffmpeg.exe");
-    let mut dest_file = File::create(&dest_path)?;
 
     for i in 0..zip.len() {
         let mut file = zip.by_index(i)?;
-        if file.name().contains("ffmpeg.exe") {
+        if file.name().contains("ffmpeg.exe") || file.name().contains("ffprobe.exe") {
+            let path = Utf8Path::new(file.name());
+            let file_name = path.file_name().unwrap();
+
+            let dest_path = destination_folder.as_ref().join(file_name);
+            let mut dest_file = File::create(&dest_path)?;
             std::io::copy(&mut file, &mut dest_file)?;
-            return Ok(dest_path);
         }
     }
-    bail!("no file found")
+
+    Ok(())
 }
 
 #[cfg(target_os = "macos")]
-fn unzip(
-    _path: impl AsRef<Utf8Path>,
-    _destination_folder: impl AsRef<Utf8Path>,
-) -> Result<Utf8PathBuf> {
+fn unzip(_path: impl AsRef<Utf8Path>, _destination_folder: impl AsRef<Utf8Path>) -> Result<()> {
     bail!("Please use `brew` to install ffmpeg manually.")
 }
 
@@ -124,50 +121,69 @@ async fn download_archive(url: &str, destination: &Utf8Path) -> Result<()> {
     Ok(())
 }
 
-pub async fn download_ffmpeg(directories: &Directories) -> Result<Utf8PathBuf> {
-    let dest = directories.cache_dir().join("ffmpeg");
-    fs::create_dir_all(&dest)?;
-    if is_installed(&dest) {
-        info!("ffmpeg already installed, not doing anything.");
-        return Ok("ffmpeg".into());
+#[derive(Debug, Clone)]
+pub enum FfmpegLocation {
+    System,
+    Local(Arc<Utf8PathBuf>),
+}
+
+impl FfmpegLocation {
+    pub fn ffmpeg(&self) -> Utf8PathBuf {
+        match self {
+            FfmpegLocation::System => Utf8PathBuf::from("ffmpeg"),
+            FfmpegLocation::Local(path) => path.join("ffmpeg"),
+        }
     }
 
+    pub fn ffprobe(&self) -> Utf8PathBuf {
+        match self {
+            FfmpegLocation::System => Utf8PathBuf::from("ffprobe"),
+            FfmpegLocation::Local(path) => path.join("ffprobe"),
+        }
+    }
+}
+
+pub async fn download_ffmpeg(directories: &Directories) -> Result<FfmpegLocation> {
+    let dest = directories.cache_dir().join("ffmpeg");
+    fs::create_dir_all(&dest)?;
+
+    let existing_install = get_ffmpeg_location(&dest);
+    if let Some(location) = existing_install {
+        info!("found existing ffmpeg binaries at {location:?}");
+        return Ok(location);
+    }
+
+    info!("ffmpeg not found, downlaoding...");
     let (url, file_name) = download_url()?;
+    info!("downloading ffmpeg from {}", url);
 
     let archive_dest = dest.join(file_name);
     if !archive_dest.is_file() {
         download_archive(url, &archive_dest).await?;
     }
 
-    info!("downloaded ffmpeg archive");
-    let ffmpeg_path = unzip(&archive_dest, dest)?;
+    info!("downloaded ffmpeg archive to {archive_dest}, extracting...");
+    unzip(&archive_dest, &dest)?;
     fs::remove_file(archive_dest)?;
 
-    Ok(ffmpeg_path)
+    Ok(FfmpegLocation::Local(Arc::new(dest)))
 }
 
 #[derive(Default)]
 pub struct Ffmpeg {
     executable_path: Utf8PathBuf,
     inputs: Vec<String>,
-    output: Option<String>,
     start: Option<String>,
-    duration: Option<String>,
-    video_codec: Option<String>,
-    audio_codec: Option<String>,
-    crf: Option<String>,
-    preset: Option<String>,
     extra_args: Vec<String>,
-    video_filter: Option<String>,
     working_directory: Option<Utf8PathBuf>,
     output_file: String,
 }
 
 impl Ffmpeg {
-    pub fn new(executable: impl AsRef<Utf8Path>, output_file: String) -> Self {
+    pub fn new(location: &FfmpegLocation, output_file: String) -> Self {
         Ffmpeg {
             output_file,
-            executable_path: executable.as_ref().to_owned(),
+            executable_path: location.ffmpeg(),
             ..Default::default()
         }
     }
@@ -187,13 +203,13 @@ impl Ffmpeg {
         self
     }
 
-    pub fn duration(&mut self, duration: f64) -> &mut Self {
-        self.duration = Some(duration.to_string());
+    pub fn extra_arg(&mut self, arg: impl Into<String>) -> &mut Self {
+        self.extra_args.push(arg.into());
         self
     }
 
-    pub async fn run(&self) -> Result<()> {
-        let mut args = vec!["-hide-banner", "-loglevel", "warning"];
+    fn build_args(&self) -> Vec<&str> {
+        let mut args = vec!["-hide_banner", "-loglevel", "warning"];
 
         if let Some(start) = &self.start {
             args.push("-ss");
@@ -204,8 +220,16 @@ impl Ffmpeg {
             args.push("-i");
             args.push(input);
         }
+        for extra_arg in &self.extra_args {
+            args.push(extra_arg);
+        }
 
         args.push(&self.output_file);
+        args
+    }
+
+    pub async fn run(&self) -> Result<()> {
+        let args = self.build_args();
         let mut command = Command::new(self.executable_path.as_str());
         command.args(args);