+++ backend/src/service/local_video.rs
@@ -7,17 +7,20 @@ use tracing::{debug, info};
 use url::Url;
 use walkdir::WalkDir;
 
+use super::commands::ffmpeg::FfmpegLocation;
 use super::directories::Directories;
 use crate::data::database::{Database, DbVideo, LocalVideoSource, LocalVideoWithMarkers};
 use crate::server::handlers::AppState;
 use crate::service::commands::{ffprobe, YtDlp, YtDlpOptions};
 use crate::service::directories::FolderType;
+use crate::service::preview_image::PreviewGenerator;
 use crate::util::generate_id;
 use crate::Result;
 
 pub struct VideoService {
     database: Database,
     directories: Directories,
+    ffmpeg_location: FfmpegLocation,
 }
 
 impl From<Arc<AppState>> for VideoService {
@@ -25,6 +28,7 @@ impl From<Arc<AppState>> for VideoService {
         VideoService {
             database: value.database.clone(),
             directories: value.directories.clone(),
+            ffmpeg_location: value.ffmpeg_location.clone(),
         }
     }
 }
@@ -60,15 +64,24 @@ impl VideoService {
                     videos.push(video);
                 } else {
                     let interactive = path.with_extension("funscript").is_file();
-                    let ffprobe = ffprobe(&path).await?;
+                    let ffprobe = ffprobe(&path, &self.ffmpeg_location).await?;
                     let duration = ffprobe.duration();
                     let id = generate_id();
+                    let preview_generator = PreviewGenerator::new(
+                        self.directories.clone(),
+                        self.ffmpeg_location.clone(),
+                    );
+                    let image_path = preview_generator
+                        .generate_preview(&id, &path, duration.map_or(0.0, |d| d / 2.0))
+                        .await?;
+
                     let video = DbVideo {
                         id,
                         file_path: path.to_string(),
                         interactive,
                         source: LocalVideoSource::Folder,
                         duration: duration.unwrap_or_default(),
+                        video_preview_image: Some(image_path.to_string()),
                     };
                     info!("inserting new video {video:#?}");
                     self.database.persist_video(video.clone()).await?;
@@ -94,13 +107,18 @@ impl VideoService {
             extract_audio: false,
             destination: FolderType::Videos,
         };
-        let result = downloader.run(&options).await?;
+        let result = downloader.run(&options, &self.ffmpeg_location).await?;
         Ok((result.generated_id, result.downloaded_file))
     }
 
     pub async fn persist_downloaded_video(&self, id: String, path: Utf8PathBuf) -> Result<DbVideo> {
-        let ffprobe = ffprobe(&path).await?;
+        let ffprobe = ffprobe(&path, &self.ffmpeg_location).await?;
         let duration = ffprobe.duration();
+        let preview_generator =
+            PreviewGenerator::new(self.directories.clone(), self.ffmpeg_location.clone());
+        let image_path = preview_generator
+            .generate_preview(&id, &path, duration.map_or(0.0, |d| d / 2.0))
+            .await?;
 
         let video = DbVideo {
             id,
@@ -108,6 +126,7 @@ impl VideoService {
             interactive: false,
             source: LocalVideoSource::Download,
             duration: duration.unwrap_or_default(),
+            video_preview_image: Some(image_path.to_string()),
         };
         info!("persisting downloaded video {video:#?}");
         self.database.persist_video(video.clone()).await?;