+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## Unreleased
-- feat: add browser notifications when downloads are finished or video generation is done
+- feat: Added browser notifications when downloads are finished or video generation is done
+- feat: Generate video and marker preview images for local videos and show them in the UI
+- feat: Added filter to local video list page
 
 ## 0.10.0
 - feat: Allow specifying weights for markers